Pull out post and delete rating functions into apicalls
